Pull messy parts of sphere construction into its own thing.
