+++ Assets/Source/Debug/Render/RenderContext.cs
@@ -370,17 +370,20 @@ namespace CreateAR.EnkluPlayer
         /// <param name="iterations">The number of iterations to use to smooth out the sphere.</param>
         public void Sphere(int iterations)
         {
-            var verts = _IcosahedronVertices;
-            var tris = _IcosahedronTriangles;
+            Vector3[] verts;
+            int[] tris;
+            GeometryHelper.GeoSphere(iterations, out verts, out tris);
 
-            if (iterations > 0)
-            {
-                for (var i = 0; i < iterations; i++)
-                {
-                    Subdivide(ref verts, ref tris);
-                }
-            }
+            RenderTriangles(ref tris, ref verts);
+        }
 
+        /// <summary>
+        /// Renders triangles.
+        /// </summary>
+        /// <param name="tris">Indices into vert array.</param>
+        /// <param name="verts">Verts.</param>
+        private void RenderTriangles(ref int[] tris, ref Vector3[] verts)
+        {
             Setup(GL.LINES);
             {
                 for (int i = 0, len = tris.Length; i < len; i += 3)
@@ -388,136 +391,20 @@ namespace CreateAR.EnkluPlayer
                     var va = verts[tris[i]].normalized;
                     var vb = verts[tris[i + 1]].normalized;
                     var vc = verts[tris[i + 2]].normalized;
-                    
+
                     var a = _current.MultiplyPoint3x4(va);
                     var b = _current.MultiplyPoint3x4(vb);
                     var c = _current.MultiplyPoint3x4(vc);
 
-                    GL.Vertex(a); GL.Vertex(b);
-                    GL.Vertex(b); GL.Vertex(c);
-                    GL.Vertex(c); GL.Vertex(a);
+                    GL.Vertex(a);
+                    GL.Vertex(b);
+                    GL.Vertex(b);
+                    GL.Vertex(c);
+                    GL.Vertex(c);
+                    GL.Vertex(a);
                 }
             }
             Teardown();
         }
-
-        /// <summary>
-        /// Subdivides all triangles.
-        /// </summary>
-        public static void Subdivide(
-            ref Vector3[] vertices,
-            ref int[] triangles)
-        {
-            // cache of midpoint indices
-            var midpointIndices = new Dictionary<string, int>();
-
-            // create lists instead...
-            List<int> indexList = new List<int>(4 * triangles.Length);
-            List<Vector3> vertexList = new List<Vector3>(vertices);
-
-            // subdivide each triangle
-            for (var i = 0; i < triangles.Length - 2; i += 3)
-            {
-                // grab indices of triangle
-                int i0 = triangles[i];
-                int i1 = triangles[i + 1];
-                int i2 = triangles[i + 2];
-
-                // calculate new indices
-                int m01 = GetMidpointIndex(midpointIndices, vertexList, i0, i1);
-                int m12 = GetMidpointIndex(midpointIndices, vertexList, i1, i2);
-                int m02 = GetMidpointIndex(midpointIndices, vertexList, i2, i0);
-
-                indexList.AddRange(
-                    new[] {
-                        i0,m01,m02,
-                        i1,m12,m01,
-                        i2,m02,m12,
-                        m02,m01,m12
-                        });
-            }
-
-            // save
-            triangles = indexList.ToArray();
-            vertices = vertexList.ToArray();
-        }
-
-        /// <summary>
-        /// Used by Subdivide method.
-        /// </summary>
-        private static int GetMidpointIndex(Dictionary<string, int> midpointIndices, List<Vector3> vertices, int i0, int i1)
-        {
-            // create a key
-            string edgeKey = string.Format("{0}_{1}", Mathf.Min(i0, i1), Mathf.Max(i0, i1));
-
-            int midpointIndex = -1;
-
-            // if there is not index already...
-            if (!midpointIndices.TryGetValue(edgeKey, out midpointIndex))
-            {
-                // grab the vertex values
-                Vector3 v0 = vertices[i0];
-                Vector3 v1 = vertices[i1];
-
-                // calculate
-                var midpoint = (v0 + v1) / 2f;
-
-                // save
-                if (vertices.Contains(midpoint))
-                {
-                    midpointIndex = vertices.IndexOf(midpoint);
-                }
-                else
-                {
-                    midpointIndex = vertices.Count;
-                    vertices.Add(midpoint);
-                }
-            }
-
-            return midpointIndex;
-        }
-
-        private static readonly Vector3[] _IcosahedronVertices =
-        {
-            new Vector3(-0.525731112119133606f, 0f, 0.850650808352039932f),
-            new Vector3(0.525731112119133606f, 0f, 0.850650808352039932f),
-            new Vector3(-0.525731112119133606f, 0f, -0.850650808352039932f),
-
-            new Vector3(0.525731112119133606f, 0f, -0.850650808352039932f),
-            new Vector3(0f, 0.850650808352039932f, 0.525731112119133606f),
-            new Vector3(0f, 0.850650808352039932f, -0.525731112119133606f),
-
-            new Vector3(0f, -0.850650808352039932f, 0.525731112119133606f),
-            new Vector3(0f, -0.850650808352039932f, -0.525731112119133606f),
-            new Vector3(0.850650808352039932f, 0.525731112119133606f, 0f),
-
-            new Vector3(-0.850650808352039932f, 0.525731112119133606f, 0f),
-            new Vector3(0.850650808352039932f, -0.525731112119133606f, 0f),
-            new Vector3(-0.850650808352039932f, -0.525731112119133606f, 0f)
-        };
-
-        private static readonly int[] _IcosahedronTriangles =
-        {
-            1,4,0,
-            4,9,0,
-            4,5,9,
-            8,5,4,
-            1,8,4,
-            1,10,8,
-            10,3,8,
-            8,3,5,
-            3,2,5,
-            3,7,2,
-            3,10,7,
-            10,6,7,
-            6,11,7,
-            6,0,11,
-            6,1,0,
-            10,1,6,
-            11,0,9,
-            2,11,9,
-            5,2,9,
-            11,2,7
-        };
     }
 }
